refactor(breaking): Leverage traits for `from_gguf()` / `from_ggml()`
This introduces a slight breaking change, in that using these `from_gguf()` / `from_ggml()` methods now requires importing the trait into scope.

The methods drop the `pub` prefix as they inherit `pub` from the trait requirement itself.

The purpose of this trait is to not require each model to duplicate the structs to params mapping helper method. Instead that can be centralized.