tests: Skip encoder test case for special tokens
When the encoder adds special tokens (`true`), it also runs the decoder with `false` to not skip processing special tokens.

There is a mismatch in the output between HF and GGUF tokenizers during the decode, where the GGUF is missing an initial `<s> `.

Advice is to skip this test case for now.