"Fix" sequence breaker tokenization
Most tokenizers encode punctuation tokens differently depending on where they occur in the input, and which tokens surround them. With the default sequence breakers, the appropriate encoding usually corresponds to the encoding produced when the token occurs after a word, rather than by itself. To emulate this, prefix the token with "a" before encoding, and extract the final token of the result.

See https://github.com/LostRuins/koboldcpp/pull/982 for a correct solution to this problem.