Target generic CPUs when building Python package
This change adjusts the configuration for Python package builds to
target "generic" CPUs rather than the native CPU of this build
system. This should produce wheels with maximum CPU compatibility.

Prior to this change, CPUs lacking features found on the build system
would fail to run the Python package properly and fail with an
"Illegal Instruction" error message.